export-sig/requirements.txt
fastapi==0.75.0
uvicorn==0.18.3

ultralytics==8.1.0
opencv-python-headless==4.8.1.78
pillow==9.5.0
numpy==1.24.3
mysql-connector-python

torch==2.1.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
